+++ codec/mod.ts
@@ -1,3 +1,8 @@
+/**
+ * Encoding and decoding utilities for Deno KV keys and values.
+ * @module
+ */
+
 export * from "./types.ts";
 export * from "./KeyCodec.ts";
 export * from "./KeySerialization.ts";